pybt/requirements.txt
pyobjc==2.5.1
pyobjc-core==2.5.1
